+++ lib/zip_kit/streamer.rb
@@ -115,34 +115,6 @@ class ZipKit::Streamer
     archive.close
   end
 
-  # Creates a new Streamer wrapped in a yielding enumerator. The enumerator can be read from using `each`,
-  # and the creation of the ZIP is in lockstep with the caller calling `each` on the returned
-  # output enumerator object. This can be used when the calling program wants to stream the
-  # output of the ZIP archive and throttle that output, or split it into chunks, or use it
-  # as a generator.
-  #
-  # For example:
-  #
-  #     # The block given to {output_enum} won't be executed immediately - rather it
-  #     # will only start to execute when the caller starts to read from the output
-  #     # by calling `each`
-  #     body = ::ZipKit::Streamer.output_enum(writer: CustomWriter) do |zip|
-  #       streamer.add_stored_entry(filename: 'large.tif', size: 1289894, crc32: 198210)
-  #       streamer << large_file.read(1024*1024) until large_file.eof?
-  #       ...
-  #     end
-  #
-  #     body.each do |bin_string|
-  #       # Send the output somewhere, buffer it in a file etc.
-  #       ...
-  #     end
-  #
-  # @param kwargs_for_new [Hash] keyword arguments for {Streamer.new}
-  # @return [ZipKit::OutputEnumerator] the enumerator you can read bytestrings of the ZIP from by calling `each`
-  def self.output_enum(**kwargs_for_new, &zip_streamer_block)
-    ZipKit::OutputEnumerator.new(**kwargs_for_new, &zip_streamer_block)
-  end
-
   # Creates a new Streamer on top of the given IO-ish object.
   #
   # @param writable[#<<] the destination IO for the ZIP. Anything that responds to `<<` can be used.